# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Persona Chat API","version":"1.0.0"},"paths":{"/auth/signup":{"post":{"tags":["Auth"],"summary":"Signup","operationId":"signup_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/verify":{"get":{"tags":["Auth"],"summary":"Verify Token","operationId":"verify_token_auth_verify_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/auth/reset-password":{"post":{"tags":["Auth"],"summary":"Reset Password","description":"Helper endpoint to reset password in case of hashing issues","operationId":"reset_password_auth_reset_password_post","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","title":"Email"}},{"name":"new_password","in":"query","required":true,"schema":{"type":"string","title":"New Password"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Auth"],"summary":"Get Current User Info","description":"Get current user information including quiz status","operationId":"get_current_user_info_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/auth/logout":{"post":{"tags":["Auth"],"summary":"Logout","description":"Logout endpoint - clears client-side token","operationId":"logout_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/quiz/":{"get":{"tags":["Quiz"],"summary":"Get Quiz","description":"Get quiz questions - requires authentication","operationId":"get_quiz_quiz__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/quiz/submit":{"post":{"tags":["Quiz"],"summary":"Submit Quiz","operationId":"submit_quiz_quiz_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSubmitRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizResultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/quiz/my-persona":{"get":{"tags":["Quiz"],"summary":"Get My Persona","description":"Get the current user's persona","operationId":"get_my_persona_quiz_my_persona_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/chat/":{"post":{"tags":["Chat"],"summary":"Chat","operationId":"chat_chat__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/chat/history":{"get":{"tags":["Chat"],"summary":"Get Chat History","description":"Get chat history for the authenticated user","operationId":"get_chat_history_chat_history_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"persona","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by persona name","title":"Persona"},"description":"Filter by persona name"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","description":"Number of messages to return","default":50,"title":"Limit"},"description":"Number of messages to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatHistoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Chat"],"summary":"Clear Chat History","description":"Clear chat history for the authenticated user","operationId":"clear_chat_history_chat_history_delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"persona","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Clear history for specific persona","title":"Persona"},"description":"Clear history for specific persona"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/personas/":{"get":{"tags":["Personas"],"summary":"List Personas","description":"List all personas - requires authentication","operationId":"list_personas_personas__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PersonaResponse"},"type":"array","title":"Response List Personas Personas  Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Personas"],"summary":"Create Persona","description":"Create a new persona - requires authentication","operationId":"create_persona_personas__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/personas/{name}":{"get":{"tags":["Personas"],"summary":"Get Persona","description":"Get specific persona by name - requires authentication","operationId":"get_persona_personas__name__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Personas"],"summary":"Update Persona","description":"Update a persona - requires authentication","operationId":"update_persona_personas__name__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Personas"],"summary":"Delete Persona","description":"Delete a persona - requires authentication","operationId":"delete_persona_personas__name__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatHistoryResponse":{"properties":{"user_id":{"type":"integer","title":"User Id"},"persona":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Persona"},"count":{"type":"integer","title":"Count"},"messages":{"items":{"$ref":"#/components/schemas/ChatMessageResponse"},"type":"array","title":"Messages"}},"type":"object","required":["user_id","persona","count","messages"],"title":"ChatHistoryResponse"},"ChatMessageResponse":{"properties":{"id":{"type":"integer","title":"Id"},"persona":{"type":"string","title":"Persona"},"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"created_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created At"}},"type":"object","required":["id","persona","role","content","created_at"],"title":"ChatMessageResponse"},"ChatRequest":{"properties":{"message":{"type":"string","minLength":1,"title":"Message","description":"Message to send"},"persona":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Persona","description":"Persona name to chat with (optional, will use user's persona if not provided)"}},"type":"object","required":["message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"reply":{"type":"string","title":"Reply","description":"Assistant's reply"}},"type":"object","required":["reply"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"User email address"},"password":{"type":"string","title":"Password","description":"User password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"PersonaCreate":{"properties":{"name":{"type":"string","title":"Name","description":"Persona name"},"config":{"additionalProperties":true,"type":"object","title":"Config","description":"Persona configuration"}},"type":"object","required":["name","config"],"title":"PersonaCreate"},"PersonaResponse":{"properties":{"name":{"type":"string","title":"Name","description":"Persona name"},"config":{"additionalProperties":true,"type":"object","title":"Config","description":"Persona configuration"},"id":{"type":"integer","title":"Id","description":"Persona ID"}},"type":"object","required":["name","config","id"],"title":"PersonaResponse"},"QuizResultResponse":{"properties":{"persona":{"type":"string","title":"Persona","description":"Assigned persona name"}},"type":"object","required":["persona"],"title":"QuizResultResponse"},"QuizSubmitRequest":{"properties":{"answers":{"items":{"type":"string"},"type":"array","minItems":5,"title":"Answers","description":"List of answers (A-E)"}},"type":"object","required":["answers"],"title":"QuizSubmitRequest"},"SignupRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"User email address"},"password":{"type":"string","minLength":6,"title":"Password","description":"Password (min 6 characters)"}},"type":"object","required":["email","password"],"title":"SignupRequest"},"TokenResponse":{"properties":{"token":{"type":"string","title":"Token","description":"JWT access token"},"token_type":{"type":"string","title":"Token Type","description":"Token type","default":"bearer"}},"type":"object","required":["token"],"title":"TokenResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}